Coran Wyvern Dialog Fix for BGEE
This change False() outs a new script block in CORAN.BCS for BGEE that
detects killing a Wyvern for Coran’s quest.  This block was causing
Coran to spam a dialog line after killing a non-quest wyvern.  Thanks
to Mike1072 for help with improving code for the fix.

http://forums.gibberlings3.net/index.php?showtopic=26447

Also, includes a fix of a typo in previous PinkXan portrait fix for
BGEE.