Colquetle Family Amulet Fix
A bug report from the BGEE forums: There are two Colquetle Family Amulets in AR4300.  This fix patches AR4300.BCS to deactivate the script block that creates the second amulet.
 
According to Balquo:
Originally the area script for this area (ar4300) created the dead bodies and the amulet but was later removed and we added them differently. We removed the script from the area; but it still exists. If a mod were to add this back then you could encounter the problem you are having.
 
http://forum.baldursgate.com/discussion/comment/539816/#Comment_539816