Updating x#ict3.d
Updating Sharteel’s & Xan’s interjections with Coran.  Xan’s
interjection was ending with a question, “So are you as jittery this
cold-footed, long-faced fellow?" with responses “Yea” and “Nea”.
However, the question being answered was Coran’s original query to go
hunt wyverns.  Therefore, the player could easily answer “wrong” and
Coran would not join.

http://forum.baldursgate.com/discussion/comment/542149/#Comment_542149

I noted that the original interjection at x#xaint.d had another line
for Coran, which is still in the x#ict3.tra,

@626  = ~Do we hunt together?~

This is a much better line for Coran to say before the player replies,
as it is re-asking the question.