Update x#kiint.d
Line 386: Change SOLVED_JOURNAL entry to JOURNAL for "Teven Attacks"
journal entry. There is no open quest for this, so it seems better as
an INFO entry.

Line 601: Change SOLVED_JOURNAL entry to JOURNAL for "Kivan and Tazok"
entry.  There is a previous Kivan and Tazok journal entry from the
vanilla game, but it is not opened as a quest, but rather as an INFO
entry.