Fix Safana-Dorn dialog issue - BGEE ONLY
In BGEE SAFANJ.DLG, State 7 starts the banter with the following
trigger:
Global("DORN_SAFANA_DIALOG1","GLOBAL",1)
It is missing an action to increment the variable once the banter has
fired.  This can block mod dialogs for Safana.  Adding an action to fix
the issue:
SetGlobal("DORN_SAFANA_DIALOG1","GLOBAL",2)