Safana Darkwalk PID Fix
Safana's response to PID option "Do you have any advice, Safana?"
References Narlen Darkwalk, even in Chapter 1.

On line 2458 of x@pcinit.d, the dialog trigger contains:

OR(2)
	GlobalGT("NarlenMove","GLOBAL",0)
	GlobalLT("NarlenMove","GLOBAL",10)

which will always evaluate to true.  The intention was presumably to
trigger on NarlenMove values of 1 to 9.

Fix: Remove the OR(2) from the triggers.

http://gibberlings3.net/forums/index.php?showtopic=26925