Fixing conditions on Xzar dialog
The XZNI1.6 dialog in xzint.d has an issue with the conditions.  If
Montanan is in the party, alive, but disabled (e.g., held, panicked,
etc..), the dialog cannot continue.  Updating the second transition
condition to be “OR(3) !InParty("montaron") !InMyArea("montaron")
StateCheck("montaron",CD_STATE_NOTVALID)” which is the opposite of the
first transition, ensuring one branch can fire.

Updating the second transition’s conditions to be opposite the first

http://gibberlings3.net/forums/index.php?showtopic=26514&p=231187