[BGEE] Fix Area Scripts for AR2400 and AR2900
The area scripts for AR2400 and AR2900 in BGEE contain script blocks
that depend on the variable TazokSays, do not increment the variable.
So, both blocks will continuously fire if TazokSays = 1. This has the
effect of preventing any blocks appended to the ends of these BCS files
from executing after the party has entered AR1900 (which sets TazokSays
= 1).

This fix adds a condition to the TazokSays blocks, checking for Raiken
(AR2400) or Teven (AR2900) to Exist().  This should allow them to
return FALSE once Raiken or Teven has done the DestroySelf() action.