Fix: Minsc stutter bug at Lighthouse
In x#miint2.baf, a dialogue referring to the Lighthouse Area is
triggered. But in x#miint.d, the dialogue trigger includes the
lighthouse-area-check, which leads to a stutter in case Minsc couldn't
talk in the area.

Fix: adding the AreaCheck to the baf (trigger block), and remove it
from the .d, as per Jastey's recommendation.

http://gibberlings3.net/forums/index.php?showtopic=27582&p=238536