Fix Gatekeeper dialog if CHARNAME has been to the Friendly Arm Inn
If you attempt to return to Candlekeep, and speak to the Gatekeeper,
the dialog option "I could really use your help. I have no home, no
guardian, and I need to get to the Friendly Arm Inn" appears even if
you have already been to the FAI.

Adding a check for variable "EnteredArmInn" = 0 on that dialog option.